:truck: Move some code to internal directory
