+++ plugins/csl-node-editor/src/internal/Connection.hpp
