+++ plugins/csl-node-editor/src/internal/NodeGraph.cpp
@@ -7,7 +7,7 @@
 
 #include "NodeGraph.hpp"
 
-#include "Node.hpp"
+#include "../Node.hpp"
 #include "logger.hpp"
 
 namespace csl::nodeeditor {
